requirements.txt
# Voice Assistant Requirements
# Install with: pip install -r requirements.txt
#
# NOTE: rvc-python is NOT included here — it requires Microsoft C++ Build Tools
# to compile on Windows and RVC is disabled by default (RVC_ENABLE = False in config.py).
# If you want RVC voice conversion, install Build Tools first:
# https://visualstudio.microsoft.com/visual-cpp-build-tools/
# Then run: pip install rvc-python

# Anthropic / LLM
anthropic

# Audio I/O
pyaudio
soundfile

# ML / AI Models
torch
numpy
openai-whisper        # imported as: import whisper
silero-vad            # imported as: from silero_vad import ...
openwakeword

# TTS
kokoro

# Web / HTTP
httpx
requests

# Spotify integration
spotipy

# Utilities
python-dotenv         # imported as: from dotenv import load_dotenv
pydantic
